+++ Buildlease/Buildlease/ClientApp/src/LOGIC.tsx
@@ -1,4 +1,4 @@
-import CategoryFullView from "./components/views/CategoryFullView";
+import CategoryFullView from "./components/views_tmp/CategoryFullView";
 import Globals from "./Globals";
 
 const LOGIC = {
