+++ Buildlease/Buildlease/ClientApp/src/components/catalog/Item.tsx
@@ -3,7 +3,7 @@ import { Button, message } from 'antd';
 import { useState } from 'react';
 import API from '../../API';
 import PATH from '../../PATH';
-import ProductView from '../views/ProductView';
+import ProductView from '../views_tmp/ProductView';
 
 interface ItemProps {
     ProductView: ProductView,
