+++ Buildlease/Buildlease/ClientApp/src/components/views_tmp/CategoryFilterView.tsx
@@ -1,0 +1,10 @@
+export default interface CategoryFilterView {
+  Id: number,
+  Name: string,
+
+  UnitName: string | null,
+  MinValue: number | null,
+  MaxValue: number | null,
+  
+  Values: string[] | null,
+}
