+++ Buildlease/Buildlease/ClientApp/src/components/views_tmp/CategoryFullView.tsx
@@ -1,0 +1,7 @@
+export default interface CategoryFullView {
+  Id: number,
+  ParentId: number,
+  Name: string,
+  ProductCount: number,
+  DefaultImagePath: string,
+}
