+++ Buildlease/Buildlease/ClientApp/src/components/views_tmp/CategoryView.tsx
@@ -1,0 +1,4 @@
+export default interface CategoryView {
+  Id: number,
+  Name: string,
+}
