+++ Buildlease/Buildlease/ClientApp/src/components/views_tmp/ProductView.tsx
@@ -1,0 +1,10 @@
+export default interface ProductView {
+  Id: number,
+  Name: string,
+  ShortDescription: string,
+  ImagePath: string,
+  TotalCount: number,
+  AvailableCount: number,  
+  Price: number | null,
+  AlreadyInCart: boolean,
+}
